fix reg bug

+++ src/com/imei/app/entity/User.java
@@ -5,16 +5,19 @@ import java.util.Date;
 import org.omg.CosNaming.NamingContextExtPackage.StringNameHelper;
 
 public class User {
-	private long id;
+	private int id;
 	private String phoneNum;
 	private String pwd;
 	private String nickName;
 	private Date regDate;
 	private String city;
-	public long getId() {
+	
+	
+	
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getPhoneNum() {
